Fixed a couple of bugs from refactoring.
